cargo-bin: rename downloads to not conflict
When changes were added to support a new cargo version, I continued
to see the binaries from the cargo version be installed.  After
some searching I discovered that the files downloaded for various
cargo versions are named "cargo-nightly-<triple>.tar.gz" within
uniquely named folders.

Yocto does not take into consideration the names of directories
prior to the file when populating the downloads directly so
the cargo-bin recipe was not actually downloading new files
to the downloads directory for the new version of cargo.

To fix this, we use the "downloadname" option available for
SRC_URI to provide a name "$PN-$PR-$triple.tar.gz" that will
be unique for each combination of those variables.

Signed-off-by: Paul Osborne <paul.osborne@smartthings.com>