cargo: use matching rust release for version
To date, the cargo releases that go along with each rust release
are nightly and versioned, usually, just with a date.  Some versions,
however, like 1.14.0 do not have an easily extracted version.

Instead of trying to deal with that, we are now just using the rust
version that these versions of cargo were released with (we also
include a note about the parsed version).

Signed-off-by: Paul Osborne <osbpau@gmail.com>