Exclude WORKDIR from recipe stamps, allowing caching
During install of rust-bin and cargo-bin binaries are placed into ${S},
which itself builds on ${WORKDIR}. Normally ${WORKDIR} is ignored for
sstate cache but it is not being ignored properly in rust-bin and
cargo-bin for two different reasons:

In rust-bin, a dependency is being created on the value of ${S} instead
of creating a transitive dependency on ${S} -> ${WORKDIR} and then
ignoring ${WORKDIR}. It's not clear why exactly this is the case, but it
was confirmed by looking at the output of the stamps for the rust-bin
recipe. The fix is to explicitly exclude ${WORKDIR} from the creation of
${S} with a `vardepsexclude`

In cargo-bin the issue is more straight-forward: the ${S} value was
being explicitly assembled in Python code, so Bitbake never sees that
${S} is formed from ${WORKDIR}. Adding explicit `vardeps` and
`vardepsexcludes` in the Python section fixes this.

Signed-off-by: Nick Stevens <nick@bitcurry.com>
(cherry picked from commit 30ff1e7384a16c4286ec44447f44a3ad9dacfc93)