Eliminate "rust_target[vardeps]" taskhash mismatches
Yocto adds a vardep to rust_target on the string "rust_target[vardeps]"
which then causes taskhash mismatches because the string contents of
that variable changes each time the code is built. We don't actually
care about the "rust_target[vardeps]" string, we care about the vardeps
that string specifies (*_ARCH/*_OS). This commit adds a vardepsexclude
for "rust_target[vardeps]".

Signed-off-by: Nick Stevens <nick@bitcurry.com>