Fix valid `armv7` targets using `arm`
The check for the ability to use the armv7 target in rust-common is a
bit naive - it assumes that all armv7 chips will have the `armv7a`
TUNE_FEATURE. However, there are now additional armv7 tunes that are
still valid armv7 architectures (such as the armv7ve with virtualization
extensions).

Simple change to make the check for armv7 compatability look for tune
features that start with `armv7*` instead of manually specifying every
permutation.

Signed-off-by: Nick Stevens <nick@bitcurry.com>