fix: double-evaluation of escaped variables in wrappers
The existing implementation of the wrapper scripts (cc-wrapper,
linker-wrapper, et. al.) was passing any additional arguments to the
compiler by using just $@. This is different than "$@" - while it
seems like "$@" would put all arguments into a single, quoted expansion,
it actually has a special meaning that passes arguments as-is to the
underlying program being called.

This was discovered when an additional argument to the cc-wrapper
script, which was properly escaped initially, caused a failure to build
using meta-rust-bin because escaped quotes were being evaluated twice.
It's easiest to see in an example.

The Rust crate being compiled calls out to the C compiler, i.e. our
script, as follows:

    cc-wrapper.sh "MY_ARG=\"value with spaces\""

The initial call is properly escaped - i.e. cc-wrapper sees the arg as
MY_ARG="value with spaces", a single argument. However, previously using
$@ this was expanded again when calling the compiler:

    gcc MY_ARG="value with spaces"

In this case the arguments to gcc are passed as

    $1: MY_ARG="value
    $2: with
    $3: spaces"

Using "$@" instead avoids the extra parsing, passing the argument to
gcc as intended.

Signed-off-by: Nick Stevens <nick.stevens@smartthings.com>