Fix use in multi-config builds that don't target glibc
Building for a BeaglePlay with meta-ti yields a multi-config build,
where some of the images built are for the regular (glibc) based Linux,
and some images are for bare-metal environment.

In that bare-metal environment, TARGET_OS is set to "eabi":

    meta/conf/distro/include/tclibc-baremetal.inc:TARGET_OS:arm = "eabi"

This makes meta-rust-bin abort fatally.

This patch proposes a more graceful handling by skipping such recipes,
instead of aborting the entire build, including the images where
TARGET_OS is set to Linux. This also appears to be inline with how other
recipes in oe-core handle this.